% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{cd2030.pooled-package}
\alias{cd2030.pooled}
\alias{cd2030.pooled-package}
\title{cd2030.pooled: The Countdown to 2030 Pooled App}
\description{
The Countdown to 2030 Pooled Shiny app: build one file from several countries' saved datasets, then explore it, compare countries and export tables. Run it with run_app().
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/damurka/cd2030.pooled}
  \item Report bugs at \url{https://github.com/damurka/cd2030.pooled/issues}
}

}
\author{
\strong{Maintainer}: David Kariuki \email{hello@damurka.com} (\href{https://orcid.org/0009-0003-6159-1107}{ORCID})

Other contributors:
\itemize{
  \item African Population and Health Research Center (APHRC) [copyright holder, funder]
}

}
\keyword{internal}
